+++ plugins/mentionall.js
@@ -1,0 +1,57 @@
+let WhatsAlexa = require('../events');
+let {MessageType} = require('@adiwajshing/baileys');
+let Language = require('../language');
+let TLang = Language.getString('tagall');
+let Lang = Language.getString('admin');
+let td = Config.WORKTYPE == 'public' ? false : true
+
+ async function checkUsAdmin(message, user = message.data.participant) {
+    var grup = await message.client.groupMetadata(message.jid);
+    var sonuc = grup['participants'].map((member) => {     
+        if (member.jid.split("@")[0] == user.split("@")[0] && member.isAdmin) return true; else; return false;
+    });
+    return sonuc.includes(true);
+}
+async function checkImAdmin(message, user = message.client.user.jid) {
+    var grup = await message.client.groupMetadata(message.jid);
+    var sonuc = grup['participants'].map((member) => {
+        if (member.id.split('@')[0] === user.split('@')[0] && member.isAdmin) return true; else; return false;
+    });
+    return sonuc.includes(true);
+}
+
+WhatsAlexa.addCommand({pattern: 'mentionall', fromMe: td, desc: TLang.TAGALL_DESC}, (async (message, match) => {
+
+    var im = await checkImAdmin(message);
+    var us = await checkUsAdmin(message);
+    if (!us) return await message.client.sendMessage(message.jid,Lang.USER_NOT_ADMIN ,MessageType.text);
+    if (!im) return await message.client.sendMessage(message.jid,Lang.IM_NOT_ADMIN,MessageType.text);
+
+    grup = await message.client.groupMetadata(message.jid);
+    var jids = [];
+    mesaj = '';
+    grup['participants'].map(
+        async (uye) => {
+            mesaj += '▫@' + uye.id.split('@')[0] + ' ';
+            jids.push(uye.id.replace('c.us', 's.whatsapp.net'));
+        }
+    );
+    await message.client.sendMessage(message.jid,mesaj, MessageType.extendedText, {contextInfo: {mentionedJid: jids}, previewType: 0})
+}));
+
+WhatsAlexa.addCommand({pattern: 'mentionall', fromMe: true, dontAddCommandList: true}, (async (message, match) => {
+
+    var im = await checkImAdmin(message);
+    if (!im) return await message.client.sendMessage(message.jid,Lang.IM_NOT_ADMIN,MessageType.text);
+    
+    grup = await message.client.groupMetadata(message.jid);
+    var jids = [];
+    mesaj = '';
+    grup['participants'].map(
+        async (uye) => {
+            mesaj += '▫@' + uye.id.split('@')[0] + ' ';
+            jids.push(uye.id.replace('c.us', 's.whatsapp.net'));
+        }
+    );
+    await message.client.sendMessage(message.jid,mesaj, MessageType.extendedText, {contextInfo: {mentionedJid: jids}, previewType: 0})
+}));
