--- plugins/tagall.js
@@ -1,29 +1,0 @@
-let WhatsAlexa = require('../events');
-let {MessageType} = require('@adiwajshing/baileys');
-let Language = require('../language');
-let Lang = Language.getString('tagall');
-
-async function checkImAdmin(message, user = message.client.user.jid) {
-    var grup = await message.client.groupMetadata(message.jid);
-    var sonuc = grup['participants'].map((member) => {
-        if (member.id.split('@')[0] === user.split('@')[0] && member.isAdmin) return true; else; return false;
-    });
-    return sonuc.includes(true);
-}
-
-WhatsAlexa.addCommand({pattern: 'tagall', fromMe: true, desc: Lang.TAGALL_DESC}, (async (message, match) => {
-
-   var im = await checkImAdmin(message);
-   if (!im) return await message.client.sendMessage(message.jid,Lang.ADMİN,MessageType.text);
-
-    grup = await message.client.groupMetadata(message.jid);
-    var jids = [];
-    mesaj = '';
-    grup['participants'].map(
-        async (uye) => {
-            mesaj += '▫@' + uye.id.split('@')[0] + ' ';
-            jids.push(uye.id.replace('c.us', 's.whatsapp.net'));
-        }
-    );
-    await message.client.sendMessage(message.jid,mesaj, MessageType.extendedText, {contextInfo: {mentionedJid: jids}, previewType: 0})
-}));
